refactor: Stamp 서비스를 Command/Query 패턴으로 분리
- StampService를 StampCommandService와 StampQueryService로 분리
- 컨트롤러에서 분리된 서비스 사용하도록 수정
- 커맨드 서비스에 제네릭 createStamp 메서드 도입으로 중복 코드 제거
- 트랜잭션 관리 및 이벤트 발행 로직 개선

Resolved: #41
